+++ client/src/pages/Main/Main.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import React, { useState, useEffect, useRef, useCallback } from "react";
+import { useState, useEffect, useRef, useCallback } from "react";
 import styled from "styled-components";
 import { SortButtons, Loading } from "../../components/CommonUI";
 import { userSession } from "../../redux/slices/userSlice";
@@ -10,7 +10,6 @@ import {
   RecipeItemList,
 } from "../../components/Main";
 import axios from "axios";
-import topbtn from "../../assets/images/main/topbtn.png";
 
 const SMainLayout = styled.main`
   display: flex;
@@ -31,23 +30,6 @@ const SLoadingLayout = styled.div`
   align-items: center;
 `;
 
-const STopBtn = styled.img`
-  width: 65px;
-  position: fixed;
-  right: 45px;
-  bottom: 50px;
-  cursor: pointer;
-  @media ${({ theme }) => theme.device.tablet} {
-    width: 50px;
-    right: 20px;
-  }
-  @media ${({ theme }) => theme.device.mobile} {
-    width: 40px;
-    right: 10px;
-    bottom: 30px;
-  }
-`;
-
 const Main = () => {
   const sortValues = ["최신순", "조회순", "평점순"];
   const [mainData, setMainData] = useState<any[]>([]);
@@ -57,7 +39,6 @@ const Main = () => {
   const { sessionStatus } = useAppSelector((state) => state.user);
 
   const [page, setPage] = useState(1); // 현재 페이지
-  const [showButton, setShowButton] = useState<boolean>(false); //스크롤 탑버튼
   const [load, setLoad] = useState<boolean>(true); // 로딩 스피너 상태
   const preventRef = useRef(true); // 옵저버 중복 실행 방지
   const obsRef = useRef(null); //observer Element
@@ -87,20 +68,6 @@ const Main = () => {
       setPage((prev) => prev + 1); // 페이지 값 증가
     }
   };
-
-  useEffect(() => {
-    const ShowButtonClick = () => {
-      if (window.scrollY > 800) {
-        setShowButton(true);
-      } else {
-        setShowButton(false);
-      }
-    };
-    window.addEventListener("scroll", ShowButtonClick);
-    return () => {
-      window.removeEventListener("scroll", ShowButtonClick);
-    };
-  }, []);
 
   const getRecipePost = useCallback(
     async (mainSortBy: string) => {
@@ -152,13 +119,6 @@ const Main = () => {
     }
   };
 
-  const scrollToTop = () => {
-    window.scroll({
-      top: 0,
-      behavior: "smooth",
-    });
-  };
-
   return (
     <SMainLayout>
       <Carousel />
@@ -171,9 +131,6 @@ const Main = () => {
           {/* 옵저버 Element */}
           <div ref={obsRef}></div>
         </SLoadingLayout>
-        {showButton && (
-          <STopBtn onClick={scrollToTop} src={topbtn} alt="topbtn" />
-        )}
       </SSectionLayout>
     </SMainLayout>
   );
